packages/sardis-coinbase/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "sardis-coinbase"
version = "0.1.0"
description = "Coinbase CDP and x402 integrations for Sardis"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [{name = "Sardis Team", email = "dev@sardis.sh"}]
dependencies = [
    "httpx>=0.27.0",
    "sardis-core>=0.2.3",
]

[project.optional-dependencies]
cdp = ["cdp-sdk>=0.15.0"]

[tool.hatch.build.targets.wheel]
packages = ["src/sardis_coinbase"]
